requirements.txt
fastmcp==2.11.3
google-adk==1.12.0
python-dotenv==1.1.1
google-maps-routing==0.6.16